+++ tests/test_events.py
@@ -15,6 +15,8 @@
 
 from __future__ import annotations
 
+import pytest
+
 from watchdog.events import (
     EVENT_TYPE_CLOSED,
     EVENT_TYPE_CREATED,
@@ -32,6 +34,7 @@ from watchdog.events import (
     FileModifiedEvent,
     FileMovedEvent,
     FileOpenedEvent,
+    FileSystemEvent,
     FileSystemEventHandler,
 )
 
@@ -178,3 +181,76 @@ def test_file_system_event_handler_dispatch():
     for event in all_events:
         assert not event.is_synthetic
         handler.dispatch(event)
+
+
+@pytest.mark.parametrize(
+    "event, prez",
+    [
+        # Files
+        (
+            FileDeletedEvent("a"),
+            "<FileDeletedEvent: event_type=deleted, src_path='a', is_directory=False>",
+        ),
+        (
+            FileModifiedEvent("a"),
+            "<FileModifiedEvent: event_type=modified, src_path='a', is_directory=False>",
+        ),
+        (
+            FileCreatedEvent("a"),
+            "<FileCreatedEvent: event_type=created, src_path='a', is_directory=False>",
+        ),
+        (
+            FileMovedEvent("a", "b"),
+            "<FileMovedEvent: src_path='a', dest_path='b', is_directory=False>",
+        ),
+        (
+            FileClosedEvent("a"),
+            "<FileClosedEvent: event_type=closed, src_path='a', is_directory=False>",
+        ),
+        (
+            FileOpenedEvent("a"),
+            "<FileOpenedEvent: event_type=opened, src_path='a', is_directory=False>",
+        ),
+        # Folders
+        (
+            DirDeletedEvent("a"),
+            "<DirDeletedEvent: event_type=deleted, src_path='a', is_directory=True>",
+        ),
+        (
+            DirModifiedEvent("a"),
+            "<DirModifiedEvent: event_type=modified, src_path='a', is_directory=True>",
+        ),
+        (
+            DirCreatedEvent("a"),
+            "<DirCreatedEvent: event_type=created, src_path='a', is_directory=True>",
+        ),
+        (
+            DirMovedEvent("a", "b"),
+            "<DirMovedEvent: src_path='a', dest_path='b', is_directory=True>",
+        ),
+    ],
+)
+def test_event_repr_and_str(event: FileSystemEvent, prez: str) -> None:
+    assert repr(event) == prez
+    assert str(event) == prez
+
+
+def test_event_comparison():
+    creation1 = FileCreatedEvent("foo")
+    creation2 = FileCreatedEvent("foo")
+    creation3 = FileCreatedEvent("bar")
+    assert creation1 == creation2
+    assert creation1 != creation3
+    assert creation2 != creation3
+
+    move1 = FileMovedEvent("a", "b")
+    move2 = FileMovedEvent("a", "b")
+    move3 = FileMovedEvent("a", "c")
+    move4 = FileMovedEvent("b", "a")
+    assert creation1 != move1
+    assert move1 == move2
+    assert move1 != move3
+    assert move1 != move4
+    assert move2 != move3
+    assert move2 != move4
+    assert move3 != move4